+++ lib/swagger_docs_generator/extractor.rb
@@ -1,5 +1,7 @@
 # frozen_string_literal: true
 
+# :reek:NestedIterators
+
 module SwaggerDocsGenerator
   # # Extractor routes info
   #
@@ -11,15 +13,20 @@ module SwaggerDocsGenerator
       @routes = Rails.application.routes.routes
     end
 
+    # Extract verb to routes
     def verb
       router do |route|
         route.verb.source.to_s.delete('$' + '^')
       end
     end
 
+    # Extract path to routes and change format to parameter path
     def path
       router do |route|
-        route.path.spec.to_s.gsub('(.:format)', '.json').gsub(':id', '{id}')
+        route.path.spec.to_s.gsub('(.:format)',
+                                  '.json').gsub(/:[a-z1-9_A-Z]*/) do |word|
+          "{#{word.delete(':')}}"
+        end
       end
     end