+++ lib/swagger_docs_generator/info.rb
@@ -5,7 +5,7 @@ module SwaggerDocsGenerator
   # Name to gem
   GEM =           'swagger_docs_generator'
   # Version to gem
-  VERSION =       '0.3.3'
+  VERSION =       '0.3.4'
   # Authors contributed to gem
   AUTHORS =       ['VAILLANT Jeremy'].freeze
   # Email to ahuthors